Refactor hunger rate management by introducing HungerRateModifierResolver and removing HungerDrainRateResolver; update walk speed modifier resolution to use centralized configuration
