+++ src/Common/Models/SlotStorage.cs
@@ -47,7 +47,7 @@ namespace CarryOn.Common.Models
                     settings.AnimationSit = slotProperties["animationSit"].AsString(settings.AnimationSit);
                     settings.AnimationCrouch = slotProperties["animationCrouch"].AsString(settings.AnimationCrouch);
 
-                    if (!BlockBehaviorCarryable.DefaultWalkSpeed.TryGetValue(slot, out var speed)) speed = 0.0F;
+                    if (!CarryCode.Default.WalkSpeedModifier.TryGetValue(slot, out var speed)) speed = 0.0F;
                     settings.WalkSpeedModifier = slotProperties["walkSpeedModifier"].AsFloat(speed);
 
                     settings.WalkSpeedModifierByType = JsonHelper.ParseFloatMap(slotProperties["walkSpeedModifierByBlockType"]);
